Extract filters from Sidebar into new SearchFilters component
