+++ covid-19-support/src/components/Sidebar.vue
@@ -4,16 +4,12 @@
       <i class="fas fa-caret-down" />
     </div>
 
-    <b-list-group class="need-day-group">
-      <b-list-group-item variant="sideNav" class="need-type">
-        <h6>{{ $t('sidebar.what-do-you-need') }}</h6>
-        <b-form-select class="custom-select" :value="need" :options="needOptionGroups" @change="(opt) => $emit('need-selected', opt)" />
-      </b-list-group-item>
-      <b-list-group-item variant="sideNav">
-        <h6>{{ $t('sidebar.when-do-you-need-it') }}</h6>
-        <b-form-select :value="day" :options="dayOptions" @change="(opt) => $emit('day-selected', opt)" />
-      </b-list-group-item>
-    </b-list-group>
+    <search-filters
+      :day="day"
+      :need="need"
+      @need-selected="(opt) => $emit('need-selected', opt)"
+      @day-selected="(opt) => $emit('day-selected', opt)"
+    />
 
     <InfoPanel :infotype="'note'" :icon="'fa-info-circle'" v-if="location.currentBusiness == null || showListing">
       {{ $t('sidebar.info-about-us') }} <a href="#" @click="$bvModal.show('about-us')">{{ $t('sidebar.info-link-text') }}</a
@@ -47,17 +43,18 @@
 </template>
 
 <script>
-import { weekdays } from '../constants'
 import BusinessDetails from './BusinessDetails.vue'
 import InfoPanel from './InfoPanel.vue'
 import ResultsList from './ResultsList.vue'
+import SearchFilters from './SearchFilters.vue'
 
 export default {
   name: 'sidebar',
   components: {
     BusinessDetails,
     InfoPanel,
-    ResultsList
+    ResultsList,
+    SearchFilters
   },
   data() {
     return {
@@ -75,118 +72,11 @@ export default {
     showList: Boolean
   },
   computed: {
-    // currentBusiness() {
-    //   if (this.location == null) {
-    //     return
-    //   }
-    //   return 0 + this.filteredMarkers.length > 0 && this.location.locValue > -1 ? this.filteredMarkers[this.location.locValue] : null
-    // },
-    needOptionGroups() {
-      const categories = this.getNeedCategories().categories
-      const needOptions = [{ value: 'none', text: this.$tc('label.selectacategory', 1) }]
-      categories.forEach((category) => {
-        if (category.subcategories != undefined) {
-          const label = category.name
-          const myOptions = []
-          category.subcategories.forEach((subcategory) => {
-            const text = 'category.' + subcategory.code
-            myOptions.push({
-              text: this.$t(text),
-              value: subcategory.code
-            })
-          })
-          needOptions.push({
-            label: label,
-            options: myOptions
-          })
-        } else {
-          const text = 'category.' + category.code
-          needOptions.push({
-            text: this.$t(text),
-            value: category.code
-          })
-        }
-      })
-      return needOptions
-    },
-    needOptions() {
-      return [
-        { value: 'selectacategory', text: this.$tc('label.selectacategory', 1) },
-        { value: 'restaurant', text: this.$tc('category.restaurant', 2) },
-        { value: 'meal', text: this.$tc('category.meal', 2) },
-        { value: 'family', text: this.$tc('category.family', 2) },
-        { value: 'farm', text: this.$tc('category.farm', 2) },
-        { value: 'grocery', text: this.$tc('category.grocery', 2) },
-        { value: 'pharmacy', text: this.$tc('category.pharmacy', 1) },
-        { value: 'food_bev', text: this.$tc('category.food_bev', 2) },
-        { value: 'pet', text: this.$t('category.pet') }
-      ]
-    },
-    dayOptions() {
-      return weekdays.map((i) => ({
-        value: i.pos,
-        text: this.$t(`dayofweek.${i.day}`)
-      }))
-    },
     tabtitle() {
       return this.isFilterOpen ? this.$t('sidebar.close-panel') : this.$t('sidebar.open-panel')
     }
   },
   methods: {
-    getNeedCategories() {
-      return {
-        categories: [
-          {
-            code: 'food',
-            id: 1001,
-            name: 'Food Resources',
-            subcategories: [
-              {
-                code: 'restaurant',
-                id: 1002,
-                name: 'Restaurants'
-              },
-              {
-                code: 'meal',
-                id: 1004,
-                name: 'Free meals'
-              },
-              {
-                code: 'family',
-                id: 1006,
-                name: 'Prepared family meals'
-              },
-              {
-                code: 'food_bev',
-                id: 1003,
-                name: 'Specialty food & beverage'
-              },
-              {
-                code: 'grocery',
-                id: 1005,
-                name: 'Groceries'
-              }
-            ]
-          },
-          {
-            code: 'farm',
-            id: 1007,
-            name: 'Farms & farmers markets'
-          },
-          {
-            code: 'pharmacy',
-            id: 1008,
-            name: 'Pharmacy'
-          },
-          {
-            code: 'pet',
-            id: 1009,
-            name: 'Pet supplies'
-          }
-        ],
-        regions: ['Orange']
-      }
-    },
     closeDetails: function () {
       this.showListing = true
     },
@@ -267,24 +157,6 @@ export default {
 .list-group {
   /* border-bottom: solid 1px rgba(0, 0, 0, 0.125); */
   padding: 0;
-}
-
-.list-group-item {
-  border: none !important;
-  padding: 0 1.25rem;
-  border-bottom: none;
-
-  &.need-type {
-    margin-bottom: 8px;
-  }
-}
-
-.list-group.need-day-group {
-  padding: 1rem 0 0 0 !important;
-}
-
-.list-group-flush.need-day-group .list-group-item:first-child {
-  padding-bottom: 1rem !important;
 }
 
 .side-nav {
